+++ myst2canvas/myst2json.py
@@ -1,7 +1,6 @@
 import re
 import json
 import jupytext as jp
-import canvasapi as cv
 from jupytext.cli import jupytext
 from dateutil.parser import parse
 import urllib.parse as parseurl
@@ -353,6 +352,7 @@ def parse_question(text, question_arr):
     obj
         parsed question
     """
+    print(text)
     attrs = parse_attrs(text, title_name="question_name", desc_name="question_text")
     q_type = attrs["question_type"]
     if len(question_arr) == 0:
@@ -457,107 +457,3 @@ def parse_quiz(nb_file):
     pprint(quiz)
 
     return quiz
-
-def get_course(url, token, course_id):
-    """
-    Retrieve Course from Canvas
-
-    Parameters
-    ----------
-    url: str
-        Canvas URL
-
-    token: str
-        access token
-
-    course_id: str
-        numerical ID of course
-
-    Returns
-    -------
-    Course
-        Course object that hosts quiz
-    """
-    canvas = cv.Canvas(url, token)
-    return canvas.get_course(course_id)
-
-def edit_quiz(quiz, canvas_quiz):
-    """
-    Update an existing quiz.
-
-    Parameters
-    ----------
-    quiz: obj
-        parsed quiz data
-
-    canvas_quiz: Quiz
-        Quiz object to modify
-
-    Returns
-    -------
-    None
-    """
-    for group in quiz["groups"]:
-        if group["questions"] is []:
-            continue
-        group_attrs = [group["attrs"]]
-        if group["attrs"]["name"].lower() != "general":
-            canvas_group = canvas_quiz.create_question_group(group_attrs)
-        for question in group["questions"]:
-            if group["attrs"]["name"].lower() != "general":
-                question["quiz_group_id"] = canvas_group.id
-            canvas_quiz.create_question(question=question)
-
-def upload_quiz(quiz, url, token, course_id):
-    """
-    Upload a new quiz.
-
-    Parameters
-    ----------
-    quiz: obj
-        parsed quiz data
-
-    url: str
-        Canvas URL
-
-    token: str
-        access token
-
-    course_id: str
-        numerical ID of course
-
-    Returns
-    -------
-    None
-    """
-    canvas_quiz = get_course(url, token, course_id).create_quiz(quiz["attrs"])
-    edit_quiz(quiz, canvas_quiz)
-
-def update_quiz(quiz, url, token, course_id, quiz_id):
-    """
-    Update an existing quiz.
-
-    Parameters
-    ----------
-    quiz: obj
-        parsed quiz data
-
-    url: str
-        Canvas URL
-
-    token: str
-        access token
-
-    course_id: str
-        numerical ID of course
-
-    quiz_id: str
-        numerical ID of quiz
-
-    Returns
-    -------
-    None
-    """
-    canvas_quiz = get_course(url, token, course_id).get_quiz(quiz_id)
-    canvas_quiz = canvas_quiz.edit(quiz=quiz["attrs"])
-    edit_quiz(quiz, canvas_quiz)
